Improve layout of content in projectView repaint after item edit

+++ pasta_eln/widgetProjectTreeView.py
@@ -1,7 +1,7 @@
 """ Custom tree view on data model """
 import subprocess, os, platform, logging, shutil
 from pathlib import Path
-from PySide6.QtWidgets import QWidget, QTreeView, QAbstractItemView, QMenu # pylint: disable=no-name-in-module
+from PySide6.QtWidgets import QWidget, QTreeView, QAbstractItemView, QMenu, QMessageBox # pylint: disable=no-name-in-module
 from PySide6.QtGui import QStandardItemModel, QStandardItem  # pylint: disable=no-name-in-module
 from PySide6.QtCore import QPoint, Qt  # pylint: disable=no-name-in-module
 from .widgetProjectLeafRenderer import ProjectLeafRenderer
@@ -55,7 +55,8 @@ class TreeView(QTreeView):
       hierStack = self.currentIndex().data().split('/')
       if hierStack[-1][0]=='x':
         docType= 'x'+str(len(hierStack))
-        self.comm.backend.cwd = Path(self.comm.backend.db.getDoc(hierStack[-1])['-branch'][0]['path'])
+        docID = hierStack[-1][:-2] if hierStack[-1].endswith(' -') else hierStack[-1]
+        self.comm.backend.cwd = Path(self.comm.backend.db.getDoc(docID)['-branch'][0]['path'])
         docID = self.comm.backend.addData(docType, {'-name':'new folder'}, hierStack)
         # append item to the GUI
         item  = self.model().itemFromIndex(self.currentIndex())
@@ -69,7 +70,8 @@ class TreeView(QTreeView):
     elif menuName=='addSibling':
       hierStack= self.currentIndex().data().split('/')[:-1]
       docType= 'x'+str(len(hierStack))
-      self.comm.backend.cwd = Path(self.comm.backend.db.getDoc(hierStack[-1])['-branch'][0]['path'])
+      docID = hierStack[-1][:-2] if hierStack[-1].endswith(' -') else hierStack[-1]
+      self.comm.backend.cwd = Path(self.comm.backend.db.getDoc(docID)['-branch'][0]['path'])
       docID = self.comm.backend.addData(docType, {'-name':'new folder'}, hierStack)
       # append item to the GUI
       item  = self.model().itemFromIndex(self.currentIndex())
@@ -79,27 +81,30 @@ class TreeView(QTreeView):
       parent.appendRow(child)
       #TODO_P3 appendRow is not 100% correct: see above
     elif menuName=='del':
-      docID = self.currentIndex().data().split('/')[-1]
-      doc = self.comm.backend.db.remove(docID)
-      for branch in doc['-branch']:
-        oldPath = Path(self.comm.backend.basePath)/branch['path']
-        if oldPath.exists():
-          if (oldPath.parent/('trash_'+oldPath.name)).exists():  #ensure target does not exist
-            endText = ' was marked for deletion. Save it or its content now to some place on harddisk. It will be deleted now!!!'
-            showMessage(self, 'Warning', 'Warning! \nThe folder '+str(oldPath.parent/('trash_'+oldPath.name))+endText)
-            if (oldPath.parent/('trash_'+oldPath.name)).exists():
-              shutil.rmtree(oldPath.parent/('trash_'+oldPath.name))
-          oldPath.rename( oldPath.parent/('trash_'+oldPath.name) )
-      # go through children
-      children = self.comm.backend.db.getView('viewHierarchy/viewHierarchy', startKey=' '.join(doc['-branch'][0]['stack']+[docID,'']))
-      for line in children:
-        self.comm.backend.db.remove(line['id'])
-      # remove leaf from GUI
-      item  = self.model().itemFromIndex(self.currentIndex())
-      parent = item.parent()
-      if parent is None: #top level
-        parent = self.model().invisibleRootItem()
-      parent.removeRow(item.row())
+      ret = QMessageBox.critical(self, 'Warning', 'Are you sure you want to delete this data?',\
+                                QMessageBox.StandardButton.Yes, QMessageBox.StandardButton.No)
+      if ret==QMessageBox.StandardButton.Yes:
+        docID = self.currentIndex().data().split('/')[-1]
+        doc = self.comm.backend.db.remove(docID)
+        for branch in doc['-branch']:
+          oldPath = Path(self.comm.backend.basePath)/branch['path']
+          if oldPath.exists():
+            if (oldPath.parent/('trash_'+oldPath.name)).exists():  #ensure target does not exist
+              endText = ' was marked for deletion. Save it or its content now to some place on harddisk. It will be deleted now!!!'
+              showMessage(self, 'Warning', 'Warning! \nThe folder '+str(oldPath.parent/('trash_'+oldPath.name))+endText)
+              if (oldPath.parent/('trash_'+oldPath.name)).exists():
+                shutil.rmtree(oldPath.parent/('trash_'+oldPath.name))
+            oldPath.rename( oldPath.parent/('trash_'+oldPath.name) )
+        # go through children
+        children = self.comm.backend.db.getView('viewHierarchy/viewHierarchy', startKey=' '.join(doc['-branch'][0]['stack']+[docID,'']))
+        for line in children:
+          self.comm.backend.db.remove(line['id'])
+        # remove leaf from GUI
+        item  = self.model().itemFromIndex(self.currentIndex())
+        parent = item.parent()
+        if parent is None: #top level
+          parent = self.model().invisibleRootItem()
+        parent.removeRow(item.row())
     elif menuName=='fold':
       item = self.model().itemFromIndex(self.currentIndex())
       if item.text().endswith(' -'):
@@ -137,4 +142,6 @@ class TreeView(QTreeView):
     docID = self.currentIndex().data().split('/')[-1]
     doc   = self.comm.backend.db.getDoc(docID[:-2] if docID.endswith(' -') else docID)
     self.comm.formDoc.emit(doc)
+    item  = self.model().itemFromIndex(self.currentIndex())
+    item.emitDataChanged()  #force redraw (resizing and repainting) of this item only
     return
